Update edit and delete function

+++ delete.php
@@ -1,7 +1,10 @@
 <?php
 require("connection.php");
-$id=$_REQUEST['id'];
-mysqli_query($connect, "DELETE FROM wishlist WHERE id=$id");
-header('location: index.php');
-exit;
+if (isset($_GET['del'])) {
+	$id = $_GET['del'];
+	mysqli_query($connect, "DELETE FROM wishlist WHERE id=$id");
+	header('location: index.php');
+	exit;
+}
+
  ?>
